Revert "[RELEASE] 0.1.0.dev724 with 0.0.5 of beta (#962)"
This reverts commit 8fd8449b32f6945672bfd5a69f22a321f183fd83.
