+++ docs/css/code_select.css
@@ -23,6 +23,8 @@ div.autodoc-desc {
   margin-top: 10px;
   margin-bottom: 10px;
   padding-left: 30px;
+  border-left: 5px solid rgba(230, 230, 230);
+  font-weight: bold;
 }
 
 div.autodoc-desc p {
