+++ featurebyte/api/api_object.py
@@ -12,7 +12,6 @@ from functools import partial
 from http import HTTPStatus
 
 import pandas as pd
-from alive_progress import alive_bar
 from bson.objectid import ObjectId
 from cachetools import TTLCache, cachedmethod
 from cachetools.keys import hashkey
@@ -21,9 +20,8 @@ from pydantic import Field
 from rich.pretty import pretty_repr
 from typeguard import typechecked
 
-from featurebyte.common.env_util import get_alive_bar_additional_params
 from featurebyte.common.utils import construct_repr_string
-from featurebyte.config import Configurations, WebsocketClient
+from featurebyte.config import Configurations
 from featurebyte.exception import (
     DuplicatedRecordException,
     ObjectHasBeenSavedError,
@@ -218,38 +216,17 @@ class ApiObject(FeatureByteBaseDocumentModel):
     @classmethod
     def get(cls: Type[ApiObjectT], name: str) -> ApiObjectT:
         """
-        Retrieve the object from the persistent data store given the object's name.
-
-        This assumes that the object has been saved to the persistent data store. If the object has not been saved,
-        an exception will be raised and you should create and save the object first.
+        Retrieve object from the persistent given object name
 
         Parameters
         ----------
         name: str
-            Name of the object to retrieve.
+            Object name
 
         Returns
         -------
         ApiObjectT
-            Retrieved object with the specified name.
-
-        Examples
-        --------
-        Note that the examples below are not exhaustive.
-
-        Get an Entity object that is already saved.
-
-        >>> grocery_customer_entity = fb.Entity.get("grocerycustomer")  # doctest: +SKIP
-        Entity(name="grocerycustomer", serving_names=["GROCERYCUSTOMERGUID"])
-
-        Get an Entity object that is not saved.
-
-        >>> grocery_customer_entity = fb.Entity.get("random_entity")  # doctest: +SKIP
-        'Entity (name: "random_entity") not found. Please save the Entity object first.'
-
-        Get a Feature object that is already saved.
-
-        >>> feature = fb.Feature.get("InvoiceCount_60days")
+            Retrieved object with the specified name
         """
         return cls._get(name)
 
@@ -283,48 +260,17 @@ class ApiObject(FeatureByteBaseDocumentModel):
         cls: Type[ApiObjectT], id: ObjectId  # pylint: disable=redefined-builtin,invalid-name
     ) -> ApiObjectT:
         """
-        Retrieve the object from the persistent data store given the object's ID.
-
-        This assumes that the object has been saved to the persistent data store. If the object has not been saved,
-        an exception will be raised and you should create and save the object first.
+        Get the object from the persistent given the object ID
 
         Parameters
         ----------
         id: ObjectId
-            Object ID value.
+            Object ID value
 
         Returns
         -------
         ApiObjectT
-            ApiObject object of the given object ID.
-
-        Examples
-        --------
-        Note that the examples below are not exhaustive.
-
-        Get an Entity object that is already saved.
-
-        >>> fb.Entity.get_by_id(grocery_customer_entity_id)  # doctest: +SKIP
-        <featurebyte.api.entity.Entity at 0x7f511dda3bb0>
-        {
-          'name': 'grocerycustomer',
-          'created_at': '2023-03-22T04:08:21.668000',
-          'updated_at': '2023-03-22T04:08:22.050000',
-          'serving_names': [
-            'GROCERYCUSTOMERGUID'
-          ],
-          'catalog_name': 'grocery'
-        }
-
-        Get an Entity object that is not saved.
-
-        >>> random_object_id = ObjectId()  # doctest: +SKIP
-        >>> grocery_customer_entity = fb.Entity.get_by_id(random_object_id)  # doctest: +SKIP
-        'Entity (id: <random_object_id.id>) not found. Please save the Entity object first.'
-
-        Get a Feature object that is already saved.
-
-        >>> feature_from_id = fb.Feature.get_by_id(invoice_count_60_days_feature_id)
+            ApiObject object of the given object ID
         """
         return cls._get_by_id(id)
 
@@ -675,51 +621,6 @@ class ApiObject(FeatureByteBaseDocumentModel):
         ------
         RecordRetrievalException
             When the object cannot be found, or we have received an unexpected response status code.
-
-        Examples
-        --------
-        Most objects can have their info accessed in a similar manner. The example below is not exhaustive.
-
-        Get info of a Table object.
-
-        >>> catalog = fb.Catalog.get_or_create("grocery")  # doctest: +SKIP
-        >>> table = catalog.get_table("INVOICEITEMS")  # doctest: +SKIP
-        >>> table.info()  # doctest: +SKIP
-        {
-          'name': 'INVOICEITEMS',
-          'status': 'DRAFT',
-          'catalog_name': 'grocery',
-          'record_creation_timestamp_column': None,
-          'table_details': {
-            'database_name': 'spark_catalog',
-            'schema_name': 'GROCERY',
-            'table_name': 'INVOICEITEMS'
-          },
-          'entities': [
-            {
-              'name': 'groceryproduct',
-              'serving_names': [
-                'GROCERYPRODUCTGUID'
-              ],
-              'catalog_name': 'grocery'
-            },
-            {
-              'name': 'groceryinvoice',
-              'serving_names': [
-                'GROCERYINVOICEGUID'
-              ],
-              'catalog_name': 'grocery'
-            }
-          ],
-          'semantics': [
-            'item_id'
-          ],
-          'column_count': 8,
-          'columns_info': None,
-          'event_id_column': 'GroceryInvoiceGuid',
-          'item_id_column': 'GroceryInvoiceItemGuid',
-          'event_table_name': 'GROCERYINVOICE'
-        }
         """
         client = Configurations().get_client()
         response = client.get(url=f"{self._route}/{self.id}/info", params={"verbose": verbose})
@@ -761,56 +662,16 @@ class ApiObject(FeatureByteBaseDocumentModel):
         if create_response.status_code == HTTPStatus.CREATED:
             create_response_dict = create_response.json()
             status = create_response_dict["status"]
-            task_id = create_response_dict["id"]
-
-            def _update_progress_bar(progress_bar: Any, websocket_client: WebsocketClient) -> None:
-                """
-                Update progress bar based on websocket message
-
-                Parameters
-                ----------
-                progress_bar: Any
-                    Alive progress bar
-                websocket_client: WebsocketClient
-                    Websocket client
-                """
-                # receive message from websocket
-                while True:
-                    message = websocket_client.receive_json()
-                    if not message:
-                        break
-                    # update progress bar
-                    description = message.get("message")
-                    if description:
-                        progress_bar.text(description)
-                    percent = message.get("percent")
-                    if percent:
-                        # end of stream
-                        if percent == -1:
-                            break
-                        progress_bar(percent / 100)  # pylint: disable=not-callable
 
             # poll the task route (if the task is still running)
             task_get_response = None
-            with alive_bar(
-                manual=True,
-                title="Working...",
-                **get_alive_bar_additional_params(),
-            ) as progress_bar:
-                with Configurations().get_websocket_client(task_id=task_id) as websocket_client:
-                    while status in [TaskStatus.STARTED, TaskStatus.PENDING]:
-                        # update progress bar
-                        _update_progress_bar(progress_bar, websocket_client)
-
-                        # retrieve task status
-                        task_get_response = client.get(url=f"/task/{task_id}")
-                        if task_get_response.status_code == HTTPStatus.OK:
-                            status = task_get_response.json()["status"]
-                            time.sleep(delay)
-                        else:
-                            raise RecordRetrievalException(task_get_response)
-                progress_bar.title = "Done!"
-                progress_bar(1)  # pylint: disable=not-callable
+            while status in [TaskStatus.STARTED, TaskStatus.PENDING]:
+                task_get_response = client.get(url=f'/task/{create_response_dict["id"]}')
+                if task_get_response.status_code == HTTPStatus.OK:
+                    status = task_get_response.json()["status"]
+                    time.sleep(delay)
+                else:
+                    raise RecordRetrievalException(task_get_response)
 
             # check the task status
             if status != TaskStatus.SUCCESS:
@@ -865,45 +726,24 @@ class SavableApiObject(ApiObject):
     @typechecked
     def save(self, conflict_resolution: ConflictResolution = "raise") -> None:
         """
-        Save an object to the persistent data store.
-
-        A conflict could be triggered when the object being saved has violated a uniqueness check at the persistent
-        data store. For example, the same object ID could have been used by another record that is already stored.
-
-        In these scenarios, we can either raise an error or retrieve the object with the same name, depending on the
-        conflict resolution parameter passed in. The default behavior is to raise an error.
+        Save object to the persistent. Conflict could be triggered when the object
+        being saved has violated uniqueness check at the persistent (for example,
+        same ID has been used by another record stored at the persistent).
 
         Parameters
         ----------
         conflict_resolution: ConflictResolution
-            "raise" will raise an error when we encounter a conflict error.
-            "retrieve" will handle the conflict error by retrieving the object with the same name.
+            "raise" raises error when then counters conflict error (default)
+            "retrieve" handle conflict error by retrieving the object with the same name
 
         Raises
         ------
         ObjectHasBeenSavedError
-            If the object has been saved before.
+            If the object has been saved before
         DuplicatedRecordException
-            When a record with the same key exists at the persistent data store.
+            When record with the same key exists at the persistent
         RecordCreationException
-            When we fail to save the new object (general failure).
-
-        Examples
-        --------
-        Note that the examples below are not exhaustive.
-
-        Save a new Entity object.
-
-        >>> entity = fb.Entity(name="grocerycustomer_example", serving_names=["GROCERYCUSTOMERGUID"])  # doctest: +SKIP
-        >>> entity.save()  # doctest: +SKIP
-        None
-
-        Calling save again returns an error.
-
-        >>> entity = fb.Entity(name="grocerycustomer", serving_names=["GROCERYCUSTOMERGUID"])  # doctest: +SKIP
-        >>> entity.save()  # doctest: +SKIP
-        >>> entity.save()  # doctest: +SKIP
-        Entity (id: <entity.id>) has been saved before.
+            When fail to save the new object (general failure)
         """
         if self.saved and conflict_resolution == "raise":
             raise ObjectHasBeenSavedError(
