--- featurebyte/api/base_aggregator.py
@@ -1,173 +1,0 @@
-"""
-This module contains base aggregator related class
-"""
-from __future__ import annotations
-
-from typing import List, Optional, Type
-
-from abc import ABC, abstractmethod
-
-from featurebyte.api.feature import Feature
-from featurebyte.api.view import View
-from featurebyte.common.typing import OptionalScalar, get_or_default
-from featurebyte.enum import AggFunc, DBVarType
-from featurebyte.exception import AggregationNotSupportedForViewError
-from featurebyte.models.base import PydanticObjectId
-from featurebyte.query_graph.node import Node
-from featurebyte.query_graph.node.agg_func import AggFuncType
-
-
-class BaseAggregator(ABC):
-    """
-    BaseAggregator is the base class for aggregators in groupby
-    """
-
-    def __init__(
-        self,
-        view: View,
-        category: Optional[str],
-        entity_ids: list[PydanticObjectId],
-        keys: list[str],
-        serving_names: list[str],
-    ):
-        self.view = view
-        self.category = category
-        self.entity_ids = entity_ids
-        self.keys = keys
-        self.serving_names = serving_names
-        if not isinstance(self.view, tuple(self.supported_views)):
-            supported_views_formatted = ", ".join(
-                [view_cls.__name__ for view_cls in self.supported_views]
-            )
-            raise AggregationNotSupportedForViewError(
-                f"{self.aggregation_method_name}() is only available for {supported_views_formatted}"
-            )
-
-    @property
-    @abstractmethod
-    def supported_views(self) -> List[Type[View]]:
-        """
-        Views that support this type of aggregation
-
-        Returns
-        -------
-        List[Type[View]]
-        """
-
-    @property
-    @abstractmethod
-    def aggregation_method_name(self) -> str:
-        """
-        Aggregation method name for readable error message
-
-        Returns
-        -------
-        str
-        """
-
-    def _validate_method_and_value_column(
-        self, method: Optional[str], value_column: Optional[str]
-    ) -> None:
-        if method is None:
-            raise ValueError("method is required")
-
-        if method not in AggFunc.all():
-            raise ValueError(f"Aggregation method not supported: {method}")
-
-        if method == AggFunc.COUNT:
-            if value_column is not None:
-                raise ValueError(
-                    "Specifying value column is not allowed for COUNT aggregation;"
-                    " try setting None as the value_column"
-                )
-        else:
-            if value_column is None:
-                raise ValueError("value_column is required")
-            if value_column not in self.view.columns:
-                raise KeyError(f'Column "{value_column}" not found in {self.view}!')
-
-    @staticmethod
-    def _validate_fill_value_and_skip_fill_na(
-        fill_value: OptionalScalar, skip_fill_na: bool
-    ) -> None:
-        if fill_value is not None and skip_fill_na:
-            raise ValueError(
-                "Specifying both fill_value and skip_fill_na is not allowed;"
-                " try setting fill_value to None or skip_fill_na to False"
-            )
-
-    def _project_feature_from_groupby_node(
-        self,
-        agg_method: AggFuncType,
-        feature_name: str,
-        groupby_node: Node,
-        method: str,
-        value_column: Optional[str],
-        fill_value: OptionalScalar,
-        skip_fill_na: bool,
-    ) -> Feature:
-
-        # value_column is None for count-like aggregation method
-        input_var_type = self.view.column_var_type_map.get(value_column, DBVarType.FLOAT)  # type: ignore
-        if not agg_method.is_var_type_supported(input_var_type):
-            raise ValueError(
-                f'Aggregation method "{method}" does not support "{input_var_type}" input variable'
-            )
-
-        var_type = agg_method.derive_output_var_type(
-            input_var_type=input_var_type, category=self.category
-        )
-
-        feature = self.view._project_feature_from_node(  # pylint: disable=protected-access
-            node=groupby_node,
-            feature_name=feature_name,
-            feature_dtype=var_type,
-            entity_ids=self.entity_ids,
-        )
-        if not skip_fill_na:
-            self._fill_feature(feature, method, feature_name, fill_value)
-        return feature
-
-    def _fill_feature(
-        self,
-        feature: Feature,
-        method: str,
-        feature_name: str,
-        fill_value: OptionalScalar,
-    ) -> Feature:
-        """
-        Fill feature values as needed.
-
-        Parameters
-        ----------
-        feature: Feature
-            feature
-        method: str
-            aggregation method
-        feature_name: str
-            feature name
-        fill_value: OptionalScalar
-            value to fill
-
-        Returns
-        -------
-        Feature
-
-        Raises
-        ------
-        ValueError
-            If both fill_value and category parameters are specified
-        """
-        if fill_value is not None and self.category is not None:
-            raise ValueError("fill_value is not supported for aggregation per category")
-
-        if method in {AggFunc.COUNT, AggFunc.NA_COUNT} and self.category is None:
-            # Count features should be 0 instead of NaN when there are no records
-            value_to_fill = get_or_default(fill_value, 0)
-            feature.fillna(value_to_fill)
-            feature.name = feature_name
-        elif fill_value is not None:
-            feature.fillna(fill_value)
-            feature.name = feature_name
-
-        return feature
