+++ featurebyte/api/catalog.py
@@ -12,16 +12,14 @@ from bson import ObjectId
 from typeguard import typechecked
 
 from featurebyte.api.api_object import SavableApiObject
-from featurebyte.api.data_source import DataSource
 from featurebyte.api.entity import Entity
-from featurebyte.api.feature import Feature
+from featurebyte.api.feature import Feature, FeatureNamespace
 from featurebyte.api.feature_job_setting_analysis import FeatureJobSettingAnalysis
-from featurebyte.api.feature_list import FeatureList
+from featurebyte.api.feature_list import FeatureList, FeatureListNamespace
 from featurebyte.api.feature_store import FeatureStore
 from featurebyte.api.periodic_task import PeriodicTask
 from featurebyte.api.relationship import Relationship
 from featurebyte.api.table import Table
-from featurebyte.api.view import View
 from featurebyte.common.doc_util import FBAutoDoc
 from featurebyte.config import activate_catalog, get_active_catalog_id
 from featurebyte.exception import RecordRetrievalException
@@ -275,6 +273,7 @@ class Catalog(CatalogModel, SavableApiObject):
     def list_features(
         self,
         include_id: Optional[bool] = False,
+        feature_list_id: Optional[ObjectId] = None,
         entity: Optional[str] = None,
         data: Optional[str] = None,
     ) -> pd.DataFrame:
@@ -285,6 +284,8 @@ class Catalog(CatalogModel, SavableApiObject):
         ----------
         include_id: Optional[bool]
             Whether to include id in the list
+        feature_list_id: Optional[ObjectId] = None,
+            Include only features in specified feature list
         entity: Optional[str]
             Name of entity used to filter results
         data: Optional[str]
@@ -295,17 +296,19 @@ class Catalog(CatalogModel, SavableApiObject):
         pd.DataFrame
             Table of features
         """
-        return Feature.list(include_id=include_id, entity=entity, data=data)
+        return Feature.list_versions(
+            include_id=include_id, feature_list_id=feature_list_id, entity=entity, data=data
+        )
 
     @update_and_reset_catalog
-    def list_feature_lists(
+    def list_feature_namespaces(
         self,
         include_id: Optional[bool] = False,
         entity: Optional[str] = None,
         data: Optional[str] = None,
     ) -> pd.DataFrame:
         """
-        List saved feature lists
+        List saved feature namespaces
 
         Parameters
         ----------
@@ -319,9 +322,55 @@ class Catalog(CatalogModel, SavableApiObject):
         Returns
         -------
         pd.DataFrame
+            Table of feature namespaces
+        """
+        return FeatureNamespace.list(include_id=include_id, entity=entity, data=data)
+
+    @update_and_reset_catalog
+    def list_feature_list_namespaces(
+        self,
+        include_id: Optional[bool] = False,
+        entity: Optional[str] = None,
+        data: Optional[str] = None,
+    ) -> pd.DataFrame:
+        """
+        List saved feature list namespaces
+
+        Parameters
+        ----------
+        include_id: Optional[bool]
+            Whether to include id in the list
+        entity: Optional[str]
+            Name of entity used to filter results
+        data: Optional[str]
+            Name of table used to filter results
+
+        Returns
+        -------
+        pd.DataFrame
+            Table of feature list namespaces
+        """
+        return FeatureListNamespace.list(include_id=include_id, entity=entity, data=data)
+
+    @update_and_reset_catalog
+    def list_feature_lists(
+        self,
+        include_id: Optional[bool] = False,
+    ) -> pd.DataFrame:
+        """
+        List saved feature lists
+
+        Parameters
+        ----------
+        include_id: Optional[bool]
+            Whether to include id in the list
+
+        Returns
+        -------
+        pd.DataFrame
             Table of feature lists
         """
-        return FeatureList.list(include_id=include_id, entity=entity, data=data)
+        return FeatureList.list_versions(include_id=include_id)
 
     @update_and_reset_catalog
     def list_tables(
@@ -475,42 +524,6 @@ class Catalog(CatalogModel, SavableApiObject):
         return PeriodicTask.list(include_id=include_id)
 
     @update_and_reset_catalog
-    def get_data_source(self, feature_store_name: str) -> DataSource:
-        """
-        Get data source by name
-
-        Parameters
-        ----------
-        feature_store_name: str
-            Feature store name
-
-        Returns
-        -------
-        DataSource
-            Data source object
-        """
-        feature_store = FeatureStore.get(name=feature_store_name)
-        return feature_store.get_data_source()
-
-    @update_and_reset_catalog
-    def get_view(self, table_name: str) -> View:
-        """
-        Get view by name
-
-        Parameters
-        ----------
-        table_name: str
-            Table name
-
-        Returns
-        -------
-        View
-            View object
-        """
-        table = Table.get(name=table_name)
-        return table.get_view()  # type: ignore[no-any-return]
-
-    @update_and_reset_catalog
     def get_feature(self, name: str, version: Optional[str] = None) -> Feature:
         """
         Get a feature by name and version
@@ -528,6 +541,40 @@ class Catalog(CatalogModel, SavableApiObject):
             Feature object
         """
         return Feature.get(name=name, version=version)
+
+    @update_and_reset_catalog
+    def get_feature_namespace(self, name: str) -> FeatureNamespace:
+        """
+        Get a feature namespace by name
+
+        Parameters
+        ----------
+        name: str
+            Feature namespace name
+
+        Returns
+        -------
+        FeatureNamespace
+            Feature namespace object
+        """
+        return FeatureNamespace.get(name=name)
+
+    @update_and_reset_catalog
+    def get_feature_list_namespace(self, name: str) -> FeatureListNamespace:
+        """
+        Get a feature list namespace by name
+
+        Parameters
+        ----------
+        name: str
+            Feature list namespace name
+
+        Returns
+        -------
+        FeatureListNamespace
+            Feature list namespace object
+        """
+        return FeatureListNamespace.get(name=name)
 
     @update_and_reset_catalog
     def get_feature_list(self, name: str, version: Optional[str] = None) -> FeatureList:
