+++ featurebyte/api/source_table.py
@@ -258,7 +258,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
     def sample(
         self,
         size: int = 10,
-        seed: int = 1234,
         from_timestamp: Optional[Union[datetime, str]] = None,
         to_timestamp: Optional[Union[datetime, str]] = None,
         after_cleaning: bool = False,
@@ -271,8 +270,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
         ----------
         size: int
             Maximum number of rows to sample
-        seed: int
-            Seed to use for random sampling
         from_timestamp: Optional[datetime]
             Start of date range to sample from
         to_timestamp: Optional[datetime]
@@ -288,7 +285,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
         """
         return self.frame.sample(  # type: ignore[misc]
             size=size,
-            seed=seed,
             from_timestamp=from_timestamp,
             to_timestamp=to_timestamp,
             after_cleaning=after_cleaning,
@@ -299,7 +295,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
     def describe(
         self,
         size: int = 0,
-        seed: int = 1234,
         from_timestamp: Optional[Union[datetime, str]] = None,
         to_timestamp: Optional[Union[datetime, str]] = None,
         after_cleaning: bool = False,
@@ -312,8 +307,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
         ----------
         size: int
             Maximum number of rows to sample
-        seed: int
-            Seed to use for random sampling
         from_timestamp: Optional[datetime]
             Start of date range to sample from
         to_timestamp: Optional[datetime]
@@ -329,7 +322,6 @@ class AbstractTableData(ConstructGraphMixin, FeatureByteBaseModel, ABC):
         """
         return self.frame.describe(  # type: ignore[misc]
             size=size,
-            seed=seed,
             from_timestamp=from_timestamp,
             to_timestamp=to_timestamp,
             after_cleaning=after_cleaning,
