+++ featurebyte/core/mixin.py
@@ -244,7 +244,6 @@ class SampleMixin:
     def sample(
         self: HasExtractPrunedGraphAndNode,
         size: int = 10,
-        seed: int = 1234,
         from_timestamp: Optional[Union[datetime, str]] = None,
         to_timestamp: Optional[Union[datetime, str]] = None,
         **kwargs: Any,
@@ -256,8 +255,6 @@ class SampleMixin:
         ----------
         size: int
             Maximum number of rows to sample
-        seed: int
-            Seed to use for random sampling
         from_timestamp: Optional[datetime]
             Start of date range to sample from
         to_timestamp: Optional[datetime]
@@ -287,9 +284,7 @@ class SampleMixin:
             timestamp_column=self.timestamp_column,
         )
         client = Configurations().get_client()
-        response = client.post(
-            url=f"/feature_store/sample?size={size}&seed={seed}", json=payload.json_dict()
-        )
+        response = client.post(url=f"/feature_store/sample?size={size}", json=payload.json_dict())
         if response.status_code != HTTPStatus.OK:
             raise RecordRetrievalException(response)
         return dataframe_from_json(response.json())
@@ -298,7 +293,6 @@ class SampleMixin:
     def describe(
         self: HasExtractPrunedGraphAndNode,
         size: int = 0,
-        seed: int = 1234,
         from_timestamp: Optional[Union[datetime, str]] = None,
         to_timestamp: Optional[Union[datetime, str]] = None,
         **kwargs: Any,
@@ -310,8 +304,6 @@ class SampleMixin:
         ----------
         size: int
             Maximum number of rows to sample
-        seed: int
-            Seed to use for random sampling
         from_timestamp: Optional[datetime]
             Start of date range to sample from
         to_timestamp: Optional[datetime]
@@ -342,7 +334,7 @@ class SampleMixin:
         )
         client = Configurations().get_client()
         response = client.post(
-            url=f"/feature_store/description?size={size}&seed={seed}", json=payload.json_dict()
+            url=f"/feature_store/description?size={size}", json=payload.json_dict()
         )
         if response.status_code != HTTPStatus.OK:
             raise RecordRetrievalException(response)
