+++ featurebyte/datasets/healthcare.sql
@@ -162,6 +162,7 @@ CREATE OR REPLACE VIEW HEALTHCARE.STATEDETAILS(
   `BelowPovertyLevel`,
   `Aged65Plus`,
   `TotalPopulation`,
+  `ValidFrom`,
   `ValidTo`,
   `record_available_at`
 ) as
