+++ featurebyte/service/relationship_analysis.py
@@ -3,25 +3,23 @@ RelationshipAnalysisService class
 """
 from __future__ import annotations
 
-from typing import List, TypeVar
+from typing import List
 
 from featurebyte.models.entity import EntityModel
 
-EntityModelT = TypeVar("EntityModelT", bound=EntityModel)
 
-
-def derive_primary_entity(entities: List[EntityModelT]) -> List[EntityModelT]:
+def derive_primary_entity(entities: List[EntityModel]) -> List[EntityModel]:
     """
     Derive the primary entity from a list of entities
 
     Parameters
     ----------
-    entities: List[EntityModelT]
+    entities: List[EntityModel]
         List of entities
 
     Returns
     -------
-    List[EntityModelT]
+    List[EntityModel]
         List of main entities
     """
