+++ pyproject.toml
@@ -95,7 +95,6 @@ version = "0.1.0"
 [tool.poetry.dependencies]
 PyYAML = "^6.0"
 aiofiles = { version = "^22.1.0", optional = true }
-aioredis = "^2.0.1"
 alive-progress = "^2.4.1"
 asyncache = "^0.1.1"
 black = { version = "^22.3.0", allow-prereleases = true }
@@ -134,7 +133,6 @@ thrift-sasl = { version = "^0.4.3", optional = true }
 typeguard = "^2.13.3"
 typer = "^0.7.0"
 uvicorn = { version = "^0.18.2", extras = ["standard"], optional = true }
-websocket-client = "^1.5.1"
 wheel = "0.38.1"
 
 [tool.poetry.extras]
