+++ scripts/test-docs-setup.py
@@ -9,7 +9,7 @@ def setup() -> None:
     Setup featurebyte environment for running doctests.
     """
     # start playground
-    fb.playground(local=True, datasets=["grocery"], docs_enabled=False)
+    fb.playground(local=True, datasets=["grocery"])
 
     # create catalog
     fb.Catalog.get_or_create("grocery")
