+++ tests/integration/api/test_view_describe.py
@@ -4,12 +4,7 @@ Test API View objects describe function
 import pandas as pd
 from pandas.testing import assert_series_equal
 
-
-def _to_utc_no_offset(date):
-    """
-    Comvert timestamp to timezone naive UTC
-    """
-    return pd.to_datetime(date, utc=True).tz_localize(None)
+from tests.integration.api.feature_preview_utils import _to_utc_no_offset
 
 
 def test_event_view_describe(event_table):
